src/main.py
"""
Quantitative Stock Trading Bot - Main Entry Point
Paper trading implementation with cloud deployment support
"""

import os
import sys
from datetime import datetime
from dotenv import load_dotenv

def main():
    """Main entry point for the trading bot"""
    
    # Load environment variables
    load_dotenv()
    
    print("=" * 60)
    print("QUANTITATIVE STOCK TRADING BOT")
    print("=" * 60)
    print(f"Start Time: {datetime.now().strftime('%Y-%m-%d %H:%M:%S')}")
    print(f"Mode: {os.getenv('TRADING_MODE', 'paper').upper()}")
    print()
    
    # Validate required environment variables
    required_vars = [
        'ALPACA_PAPER_KEY_ID',
        'ALPACA_PAPER_SECRET',
    ]
    
    missing_vars = []
    for var in required_vars:
        value = os.getenv(var)
        if not value or value.startswith('your_'):
            missing_vars.append(var)
        else:
            # Show first 4 chars of keys for verification
            print(f"✓ {var}: {value[:4]}...{value[-4:]}")
    
    if missing_vars:
        print()
        print("❌ ERROR: Missing or unconfigured environment variables:")
        for var in missing_vars:
            print(f"   - {var}")
        print()
        print("Please configure these in your .env file")
        print("See .env.example for template")
        sys.exit(1)
    
    print()
    print("✓ Environment validated successfully!")
    print()
    
    # TODO: Phase 1 - Implement data layer (Alpaca API integration)
    # TODO: Phase 2 - Implement technical analysis
    # TODO: Phase 3 - Implement sentiment analysis
    # TODO: Phase 4 - Implement trading logic
    # TODO: Phase 5 - Implement backtesting
    
    print("📊 Trading logic not yet implemented")
    print("Next step: Integrate Alpaca API for data fetching")
    print()
    print("=" * 60)
    print(f"End Time: {datetime.now().strftime('%Y-%m-%d %H:%M:%S')}")
    print("=" * 60)

if __name__ == "__main__":
    main()
